fix(ui): blog -> posts
