+++ packages/ui/src/assets/posts.tsx
@@ -1,0 +1,26 @@
+import type { CSSProperties, JSX } from "react";
+
+interface Props {
+  className?: string;
+  style?: CSSProperties;
+}
+
+export const Posts = (props: Props): JSX.Element => {
+  return (
+    <svg
+      width="230"
+      height="70"
+      viewBox="0 0 230 70"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={props.className}
+      style={props.style}
+    >
+      <path
+        d="M0.544 69V0.839996H28.768C31.712 0.839996 34.4 1.448 36.832 2.664C39.328 3.88 41.472 5.54399 43.264 7.65599C45.056 9.70399 46.464 12.008 47.488 14.568C48.512 17.064 49.024 19.624 49.024 22.248C49.024 25.96 48.192 29.48 46.528 32.808C44.928 36.072 42.656 38.728 39.712 40.776C36.768 42.824 33.312 43.848 29.344 43.848H7.264V69H0.544ZM7.264 37.896H29.056C31.744 37.896 34.08 37.16 36.064 35.688C38.048 34.216 39.584 32.296 40.672 29.928C41.76 27.56 42.304 25 42.304 22.248C42.304 19.432 41.664 16.84 40.384 14.472C39.104 12.104 37.408 10.248 35.296 8.904C33.248 7.496 30.976 6.792 28.48 6.792H7.264V37.896ZM79.2933 69.96C75.7093 69.96 72.3813 69.288 69.3093 67.944C66.3013 66.536 63.6773 64.648 61.4373 62.28C59.2613 59.848 57.5653 57.096 56.3493 54.024C55.1333 50.888 54.5253 47.592 54.5253 44.136C54.5253 40.552 55.1333 37.224 56.3493 34.152C57.5653 31.016 59.2933 28.264 61.5333 25.896C63.7733 23.464 66.3973 21.576 69.4053 20.232C72.4773 18.824 75.8053 18.12 79.3893 18.12C82.9733 18.12 86.2693 18.824 89.2773 20.232C92.2853 21.576 94.9093 23.464 97.1493 25.896C99.3893 28.264 101.117 31.016 102.333 34.152C103.549 37.224 104.157 40.552 104.157 44.136C104.157 47.592 103.549 50.888 102.333 54.024C101.117 57.096 99.3893 59.848 97.1493 62.28C94.9733 64.648 92.3493 66.536 89.2773 67.944C86.2693 69.288 82.9413 69.96 79.2933 69.96ZM61.1493 44.232C61.1493 47.944 61.9493 51.336 63.5493 54.408C65.2133 57.416 67.4213 59.816 70.1733 61.608C72.9253 63.336 75.9653 64.2 79.2933 64.2C82.6213 64.2 85.6613 63.304 88.4133 61.512C91.1653 59.72 93.3733 57.288 95.0373 54.216C96.7013 51.08 97.5333 47.688 97.5333 44.04C97.5333 40.328 96.7013 36.936 95.0373 33.864C93.3733 30.792 91.1653 28.36 88.4133 26.568C85.6613 24.776 82.6213 23.88 79.2933 23.88C75.9653 23.88 72.9253 24.808 70.1733 26.664C67.4853 28.52 65.3093 30.984 63.6453 34.056C61.9813 37.064 61.1493 40.456 61.1493 44.232ZM131.491 69.96C127.395 69.96 123.587 69.288 120.067 67.944C116.547 66.6 113.507 64.552 110.947 61.8L113.635 57.192C116.451 59.816 119.267 61.736 122.083 62.952C124.963 64.104 128.003 64.68 131.203 64.68C135.107 64.68 138.275 63.912 140.707 62.376C143.139 60.776 144.355 58.504 144.355 55.56C144.355 53.576 143.747 52.072 142.531 51.048C141.379 49.96 139.683 49.096 137.443 48.456C135.267 47.752 132.643 47.016 129.571 46.248C126.115 45.288 123.203 44.296 120.835 43.272C118.531 42.184 116.771 40.84 115.555 39.24C114.403 37.576 113.827 35.432 113.827 32.808C113.827 29.544 114.627 26.856 116.227 24.744C117.891 22.568 120.131 20.936 122.947 19.848C125.827 18.696 129.027 18.12 132.547 18.12C136.387 18.12 139.779 18.728 142.723 19.944C145.667 21.16 148.067 22.856 149.923 25.032L146.755 29.448C144.963 27.4 142.819 25.896 140.323 24.936C137.891 23.912 135.171 23.4 132.163 23.4C130.115 23.4 128.163 23.688 126.307 24.264C124.451 24.776 122.915 25.672 121.699 26.952C120.547 28.168 119.971 29.864 119.971 32.04C119.971 33.832 120.419 35.24 121.315 36.264C122.211 37.224 123.555 38.056 125.347 38.76C127.139 39.4 129.347 40.104 131.971 40.872C135.747 41.896 139.043 42.952 141.859 44.04C144.675 45.064 146.851 46.408 148.387 48.072C149.923 49.736 150.691 52.072 150.691 55.08C150.691 59.752 148.931 63.4 145.411 66.024C141.891 68.648 137.251 69.96 131.491 69.96ZM185.481 66.6C184.969 66.792 184.169 67.144 183.081 67.656C181.993 68.168 180.681 68.616 179.145 69C177.609 69.384 175.945 69.576 174.153 69.576C172.297 69.576 170.537 69.224 168.873 68.52C167.273 67.816 165.993 66.76 165.033 65.352C164.073 63.88 163.593 62.088 163.593 59.976V24.168H156.681V18.984H163.593V2.088H170.121V18.984H181.641V24.168H170.121V58.152C170.249 59.944 170.889 61.288 172.041 62.184C173.257 63.08 174.633 63.528 176.169 63.528C177.961 63.528 179.593 63.24 181.065 62.664C182.537 62.024 183.433 61.576 183.753 61.32L185.481 66.6ZM210.054 69.96C205.958 69.96 202.15 69.288 198.63 67.944C195.11 66.6 192.07 64.552 189.51 61.8L192.198 57.192C195.014 59.816 197.83 61.736 200.646 62.952C203.526 64.104 206.566 64.68 209.766 64.68C213.67 64.68 216.838 63.912 219.27 62.376C221.702 60.776 222.918 58.504 222.918 55.56C222.918 53.576 222.31 52.072 221.094 51.048C219.942 49.96 218.246 49.096 216.006 48.456C213.83 47.752 211.206 47.016 208.134 46.248C204.678 45.288 201.766 44.296 199.398 43.272C197.094 42.184 195.334 40.84 194.118 39.24C192.966 37.576 192.39 35.432 192.39 32.808C192.39 29.544 193.19 26.856 194.79 24.744C196.454 22.568 198.694 20.936 201.51 19.848C204.39 18.696 207.59 18.12 211.109 18.12C214.95 18.12 218.342 18.728 221.286 19.944C224.23 21.16 226.63 22.856 228.486 25.032L225.318 29.448C223.526 27.4 221.382 25.896 218.886 24.936C216.454 23.912 213.734 23.4 210.726 23.4C208.678 23.4 206.726 23.688 204.87 24.264C203.014 24.776 201.478 25.672 200.262 26.952C199.11 28.168 198.534 29.864 198.534 32.04C198.534 33.832 198.982 35.24 199.878 36.264C200.774 37.224 202.118 38.056 203.91 38.76C205.702 39.4 207.91 40.104 210.534 40.872C214.31 41.896 217.606 42.952 220.422 44.04C223.238 45.064 225.414 46.408 226.95 48.072C228.486 49.736 229.254 52.072 229.254 55.08C229.254 59.752 227.494 63.4 223.974 66.024C220.454 68.648 215.814 69.96 210.054 69.96Z"
+        fill="none"
+        className="fill-current"
+      />
+    </svg>
+  );
+};
